+++ src/lib.rs
@@ -15,6 +15,11 @@ use std::str::FromStr;
 #[cfg(feature = "serde")]
 use serde::{Deserialize, Serialize};
 
+#[cfg(feature = "iterable")]
+mod iter;
+#[cfg(feature = "iterable")]
+pub use iter::*;
+
 macro_rules! languages_table {
     (
         $enum_name:ident, $enum_err_name:ident,
@@ -348,304 +353,6 @@ languages_table!(
     (Zu, "zu", "zul", "zul", "Zulu", "Niger–Congo"),
 );
 
-#[cfg(feature = "iterable")]
-pub mod iter {
-    use super::LanguageCode;
-
-    #[derive(Debug, Default, Clone)]
-    pub struct Iter(usize);
-
-    impl Iterator for Iter {
-        type Item = &'static LanguageCode;
-
-        fn next(&mut self) -> Option<Self::Item> {
-            if self.0 == CODES.len() {
-                return None;
-            }
-            let current = self.0;
-            self.0 += 1;
-            Some(&CODES[current])
-        }
-    }
-
-    impl Iter {
-        pub fn as_codes(self) -> Codes {
-            Codes(self)
-        }
-        pub fn as_codes_t(self) -> CodesT {
-            CodesT(self)
-        }
-        pub fn as_codes_b(self) -> CodesB {
-            CodesB(self)
-        }
-    }
-
-    #[derive(Debug, Default, Clone)]
-    pub struct Codes(Iter);
-
-    #[derive(Debug, Default, Clone)]
-    pub struct CodesT(Iter);
-
-    #[derive(Debug, Default, Clone)]
-    pub struct CodesB(Iter);
-
-    #[derive(Debug, Default, Clone)]
-    pub struct Families(usize);
-
-    impl Iterator for Codes {
-        type Item = &'static str;
-
-        fn next(&mut self) -> Option<Self::Item> {
-            self.0.next().map(|c| c.code())
-        }
-    }
-
-    impl Iterator for CodesT {
-        type Item = &'static str;
-
-        fn next(&mut self) -> Option<Self::Item> {
-            self.0.next().map(|c| c.code_t())
-        }
-    }
-
-    impl Iterator for CodesB {
-        type Item = &'static str;
-
-        fn next(&mut self) -> Option<Self::Item> {
-            self.0.next().map(|c| c.code_b())
-        }
-    }
-
-    impl Iterator for Families {
-        type Item = &'static str;
-
-        fn next(&mut self) -> Option<Self::Item> {
-            if self.0 == FAMILIES.len() {
-                return None;
-            }
-            let current = self.0;
-            self.0 += 1;
-            Some(FAMILIES[current])
-        }
-    }
-
-    pub const CODES: [LanguageCode; 184] = [
-        LanguageCode::Ab,
-        LanguageCode::Aa,
-        LanguageCode::Af,
-        LanguageCode::Ak,
-        LanguageCode::Sq,
-        LanguageCode::Am,
-        LanguageCode::Ar,
-        LanguageCode::An,
-        LanguageCode::Hy,
-        LanguageCode::As,
-        LanguageCode::Av,
-        LanguageCode::Ae,
-        LanguageCode::Ay,
-        LanguageCode::Az,
-        LanguageCode::Bm,
-        LanguageCode::Ba,
-        LanguageCode::Eu,
-        LanguageCode::Be,
-        LanguageCode::Bn,
-        LanguageCode::Bh,
-        LanguageCode::Bi,
-        LanguageCode::Bs,
-        LanguageCode::Br,
-        LanguageCode::Bg,
-        LanguageCode::My,
-        LanguageCode::Ca,
-        LanguageCode::Ch,
-        LanguageCode::Ce,
-        LanguageCode::Ny,
-        LanguageCode::Zh,
-        LanguageCode::Cv,
-        LanguageCode::Kw,
-        LanguageCode::Co,
-        LanguageCode::Cr,
-        LanguageCode::Hr,
-        LanguageCode::Cs,
-        LanguageCode::Da,
-        LanguageCode::Dv,
-        LanguageCode::Nl,
-        LanguageCode::Dz,
-        LanguageCode::En,
-        LanguageCode::Eo,
-        LanguageCode::Et,
-        LanguageCode::Ee,
-        LanguageCode::Fo,
-        LanguageCode::Fj,
-        LanguageCode::Fi,
-        LanguageCode::Fr,
-        LanguageCode::Ff,
-        LanguageCode::Gl,
-        LanguageCode::Ka,
-        LanguageCode::De,
-        LanguageCode::El,
-        LanguageCode::Gn,
-        LanguageCode::Gu,
-        LanguageCode::Ht,
-        LanguageCode::Ha,
-        LanguageCode::He,
-        LanguageCode::Hz,
-        LanguageCode::Hi,
-        LanguageCode::Ho,
-        LanguageCode::Hu,
-        LanguageCode::Ia,
-        LanguageCode::Id,
-        LanguageCode::Ie,
-        LanguageCode::Ga,
-        LanguageCode::Ig,
-        LanguageCode::Ik,
-        LanguageCode::Io,
-        LanguageCode::Is,
-        LanguageCode::It,
-        LanguageCode::Iu,
-        LanguageCode::Ja,
-        LanguageCode::Jv,
-        LanguageCode::Kl,
-        LanguageCode::Kn,
-        LanguageCode::Kr,
-        LanguageCode::Ks,
-        LanguageCode::Kk,
-        LanguageCode::Km,
-        LanguageCode::Ki,
-        LanguageCode::Rw,
-        LanguageCode::Ky,
-        LanguageCode::Kv,
-        LanguageCode::Kg,
-        LanguageCode::Ko,
-        LanguageCode::Ku,
-        LanguageCode::Kj,
-        LanguageCode::La,
-        LanguageCode::Lb,
-        LanguageCode::Lg,
-        LanguageCode::Li,
-        LanguageCode::Ln,
-        LanguageCode::Lo,
-        LanguageCode::Lt,
-        LanguageCode::Lu,
-        LanguageCode::Lv,
-        LanguageCode::Gv,
-        LanguageCode::Mk,
-        LanguageCode::Mg,
-        LanguageCode::Ms,
-        LanguageCode::Ml,
-        LanguageCode::Mt,
-        LanguageCode::Mi,
-        LanguageCode::Mr,
-        LanguageCode::Mh,
-        LanguageCode::Mn,
-        LanguageCode::Na,
-        LanguageCode::Nv,
-        LanguageCode::Nd,
-        LanguageCode::Ne,
-        LanguageCode::Ng,
-        LanguageCode::Nb,
-        LanguageCode::Nn,
-        LanguageCode::No,
-        LanguageCode::Ii,
-        LanguageCode::Nr,
-        LanguageCode::Oc,
-        LanguageCode::Oj,
-        LanguageCode::Cu,
-        LanguageCode::Om,
-        LanguageCode::Or,
-        LanguageCode::Os,
-        LanguageCode::Pa,
-        LanguageCode::Pi,
-        LanguageCode::Fa,
-        LanguageCode::Pl,
-        LanguageCode::Ps,
-        LanguageCode::Pt,
-        LanguageCode::Qu,
-        LanguageCode::Rm,
-        LanguageCode::Rn,
-        LanguageCode::Ro,
-        LanguageCode::Ru,
-        LanguageCode::Sa,
-        LanguageCode::Sc,
-        LanguageCode::Sd,
-        LanguageCode::Se,
-        LanguageCode::Sm,
-        LanguageCode::Sg,
-        LanguageCode::Sr,
-        LanguageCode::Gd,
-        LanguageCode::Sn,
-        LanguageCode::Si,
-        LanguageCode::Sk,
-        LanguageCode::Sl,
-        LanguageCode::So,
-        LanguageCode::St,
-        LanguageCode::Es,
-        LanguageCode::Su,
-        LanguageCode::Sw,
-        LanguageCode::Ss,
-        LanguageCode::Sv,
-        LanguageCode::Ta,
-        LanguageCode::Te,
-        LanguageCode::Tg,
-        LanguageCode::Th,
-        LanguageCode::Ti,
-        LanguageCode::Bo,
-        LanguageCode::Tk,
-        LanguageCode::Tl,
-        LanguageCode::Tn,
-        LanguageCode::To,
-        LanguageCode::Tr,
-        LanguageCode::Ts,
-        LanguageCode::Tt,
-        LanguageCode::Tw,
-        LanguageCode::Ty,
-        LanguageCode::Ug,
-        LanguageCode::Uk,
-        LanguageCode::Ur,
-        LanguageCode::Uz,
-        LanguageCode::Ve,
-        LanguageCode::Vi,
-        LanguageCode::Vo,
-        LanguageCode::Wa,
-        LanguageCode::Cy,
-        LanguageCode::Wo,
-        LanguageCode::Fy,
-        LanguageCode::Xh,
-        LanguageCode::Yi,
-        LanguageCode::Yo,
-        LanguageCode::Za,
-        LanguageCode::Zu,
-    ];
-
-    pub const FAMILIES: [&'static str; 26] = [
-        "Afro-Asiatic",
-        "Algonquian",
-        "Austroasiatic",
-        "Austronesian",
-        "Aymaran",
-        "Constructed",
-        "Creole",
-        "Dené–Yeniseian",
-        "Dravidian",
-        "Eskimo–Aleut",
-        "Indo-European",
-        "Japonic",
-        "Kartvelian",
-        "Koreanic",
-        "Language isolate",
-        "Mongolic",
-        "Niger–Congo",
-        "Nilo-Saharan",
-        "Northeast Caucasian",
-        "Northwest Caucasian",
-        "Quechuan",
-        "Sino-Tibetan",
-        "Tai–Kadai",
-        "Tupian",
-        "Turkic",
-        "Uralic",
-    ];
-}
-
 #[cfg(test)]
 mod tests {
     use crate::LanguageCode;
@@ -693,58 +400,5 @@ mod tests {
         assert_eq!(LanguageCode::Ae.to_string(), "Avestan");
         assert_eq!(LanguageCode::Zh.to_string(), "Chinese");
         assert_eq!(LanguageCode::Sg.to_string(), "Sango");
-    }
-
-    #[cfg(feature = "iterable")]
-    mod iter {
-        use crate::iter;
-        use crate::LanguageCode;
-
-        #[test]
-        fn language_codes() {
-            let mut codes = iter::Iter::default();
-            assert_eq!(codes.next(), Some(&LanguageCode::Ab));
-            assert_eq!(codes.next(), Some(&LanguageCode::Aa));
-        }
-
-        #[test]
-        fn all_codes() {
-            let mut codes = iter::Codes::default();
-            assert_eq!(codes.next(), Some("ab"));
-            assert_eq!(codes.next(), Some("aa"));
-
-            let mut codes = iter::Iter::default().as_codes();
-            assert_eq!(codes.next(), Some("ab"));
-            assert_eq!(codes.next(), Some("aa"));
-        }
-
-        #[test]
-        fn codes_t() {
-            let mut codes_t = iter::CodesT::default();
-            assert_eq!(codes_t.next(), Some("abk"));
-            assert_eq!(codes_t.next(), Some("aar"));
-
-            let mut codes_t = iter::Iter::default().as_codes_t();
-            assert_eq!(codes_t.next(), Some("abk"));
-            assert_eq!(codes_t.next(), Some("aar"));
-        }
-
-        #[test]
-        fn codes_b() {
-            let mut codes_b = iter::CodesB::default();
-            assert_eq!(codes_b.next(), Some("abk"));
-            assert_eq!(codes_b.next(), Some("aar"));
-
-            let mut codes_b = iter::Iter::default().as_codes_b();
-            assert_eq!(codes_b.next(), Some("abk"));
-            assert_eq!(codes_b.next(), Some("aar"));
-        }
-
-        #[test]
-        fn families() {
-            let mut families = iter::Families::default();
-            assert_eq!(families.next(), Some("Afro-Asiatic"));
-            assert_eq!(families.next(), Some("Algonquian"));
-        }
     }
 }
